backend/requirements.txt
camel-ai
fastapi
python-dotenv
uvicorn
python-multipart 